docs: fix structure/includes
When the drivers were moved to lib/common, along with some structures
that they depend on, the documentation that they generated wasn't
removed from the repository.

This commit fixes that by removing the old partials, and changing the
contents of the builder docs to correct their includes, and add sections
about some configuration structures which had not been added to the docs
when they were committed into the plugin's repository.